requirements.txt
absl-py==0.8.1
astor==0.8.0
Click==7.0
dlib==19.18.0
face-recognition==1.2.3
face-recognition-models==0.3.0
gast==0.3.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
Keras==2.3.1
Keras-Applications==1.0.8
keras-facenet==0.1a5
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.4
opencv-python==4.1.1.26
Pillow==6.2.1
pkg-resources==0.0.0
protobuf==3.10.0
PyYAML==5.1.2
scipy==1.3.2
six==1.13.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
